Add files via upload
Changed a few things from the original version. For one there are now some !pip install lines at the very beginning. That is because this notebook file was executed inside a tensorflow docker container with NVidia gpu and jupyter support (latest, not nightly). Secondly, some more code and comments were added in regards to hyperparameter tuning for the machine learning models.